More work on regexps, few function call fixes.
- Adds option parsing for regexps
- Internal function call impl now checks the ctx and replaces a
  null or undefined ctx with the global object.
- Adds function equality -- can't believe that was missing.

+++ src/str.c
@@ -19,6 +19,7 @@
 #include <stdio.h>
 #include <string.h>
 #include <stdlib.h>
+#include <stdarg.h>
 
 #include "str.h"
 
@@ -35,7 +36,7 @@ fh_str_concat(char *a, char *b)
 char *
 fh_str_slice(char *str, int start, int end)
 {
-  int size = end - start + 1;
+  size_t size = end - start + 1;
   char *new = malloc(size);
   snprintf(new, size, "%.*s", end - start, str + start);
   return new;
